% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorttools.R
\name{gv2image}
\alias{gv2image}
\title{Function makes image from graphviz code}
\usage{
gv2image(gv, file = "gv", type = "png", engine = "dot", ...)
}
\arguments{
\item{gv}{character string containing graphviz code}

\item{file}{file name for image, character string}

\item{type}{type of ('pdf', 'png', 'ps', 'raw','svg','webp') as character string}

\item{engine}{grViz engine, defaults is 'dot'}

\item{...}{parameters for rsvg_}
}
\value{
Invisible name of file created.
}
\description{
Function makes image from graphviz code
}
\author{
Jari Haukka \email{jari.haukka@helsinki.fi}
}
